requirements.txt
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
confusable-homoglyphs==3.2.0
Django==1.11.29
django-bootstrap4==2.2.0
django-fontawesome-5==1.0.18
django-material==1.7.0
django-phone-field==1.8.1
django-registration==2.4.1
django-tinymce==3.0.2
idna==2.10
importlib-metadata==1.7.0
isort==5.4.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
idna==2.10
psycopg2==2.8.5
pylint==2.6.0
python-decouple==3.3
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
<<<<<<< HEAD
toml==0.10.1
typed-ast==1.4.1
=======
>>>>>>> ba0b75ce6f5a4ffd3f343575a931f49e297552f1
urllib3==1.25.10
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.1.0
